Simulate SYMBOL TABLE using list
The symbol table (temporarily called the object table) is an important data structure created, maintained and used by compilers to keep track of the semantics of identifiers
such as storing information about names, types, scopes, etc.
In the stages that the compiler performs to convert from source code to executable code, the semantic analysis stage is one of the important stages to check the correctness of the source code, for example, checking whether a variable has been declared when used, whether the value assigned to a variable is appropriate in terms of type, etc. 
The semantic analysis stage requires this object table to track the information required by the check. In the assignment, students were asked to implement a simulation of an object register using list data structures.